Moved JSX components to .jsx files
