+++ babel.config.js
@@ -1,0 +1,13 @@
+export default {
+  presets: [
+    [
+      "next/babel",
+      {
+        "preset-env": { targets: { node: "current" }, modules: false },
+        "preset-react": { runtime: "automatic" }
+      }
+    ]
+  ],
+  plugins: ["styled-components"]
+};
+
