+++ package.json
@@ -29,7 +29,7 @@
     ]
   },
   "engines": {
-    "node": ">=17.0.1"
+    "node": ">=25.8.2"
   },
   "dependencies": {
     "@contentful/rich-text-react-renderer": "^15.16.2",
