+++ src/components/header/header.css.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import { colors } from 'constants/theme';
-import MEDIA from 'helpers/mediaTemplates';
-import LogoSvg from 'components/logo';
+import MEDIA from 'helpers/mediaTemplates.js';
+import LogoSvg from 'components/logo/logo.jsx';
 
 export const Container = styled.header`
   display: flex;
