+++ src/components/icons/flickrIcon.jsx
