--- src/components/icons/index.js
@@ -1,15 +1,0 @@
-import FivehundredpxIcon from './fivehundredpx';
-import FlickrIcon from './flickr';
-import GithubIcon from './github';
-import LinkedinIcon from './linkedin';
-import TwitterIcon from './twitter';
-import YoutubeIcon from './youtube';
-
-export {
-  FivehundredpxIcon,
-  FlickrIcon,
-  GithubIcon,
-  LinkedinIcon,
-  TwitterIcon,
-  YoutubeIcon,
-};
