+++ src/components/icons/index.jsx
@@ -1,0 +1,15 @@
+import FivehundredpxIcon from './fivehundredpxIcon.jsx';
+import FlickrIcon from './flickrIcon.jsx';
+import GithubIcon from './githubIcon.jsx';
+import LinkedinIcon from './linkedinIcon.jsx';
+import TwitterIcon from './twitterIcon.jsx';
+import YoutubeIcon from './youtubeIcon.jsx';
+
+export {
+  FivehundredpxIcon,
+  FlickrIcon,
+  GithubIcon,
+  LinkedinIcon,
+  TwitterIcon,
+  YoutubeIcon,
+};
