+++ src/components/icons/youtubeIcon.jsx
