+++ src/components/image/baseImage.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import PropTypes from 'prop-types';
-import { SImage } from './image.css';
+import { SImage } from './image.css.js';
 
 const sizes = '(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw';
